add support for 'anonymous' persons
Anonymous persons are persons that *may* disappear at any point.
They are typically created because we *know* that a person is there (eg,
we've detected a face), but that person is not yet permanently
identified.

API change: HriListener::getPersons() now returns *weak* pointers that
need to be locked before being used.